Split module graph export lookup
